--- src/layouts/default/components/SlidesBanner/types.ts
@@ -1,6 +1,0 @@
-export interface Slide {
-  text: string;
-  url: string;
-  startDate: string;
-  endDate?: string;
-}
