+++ src/types/landingPage.ts
@@ -7,3 +7,15 @@ export interface EarlyAdopter {
   title: string;
   link: string;
 }
+
+export interface Slide {
+  text: string;
+  url: string;
+  startDate: string;
+  endDate?: string;
+}
+
+export interface SlidesBannerProps {
+  bannerDuration: number;
+  slides: Slide[];
+}
